Revert "Track service data moved to get from API"
