--- Chinook/ClientModels/Request/AddTrackToPlaylistModel.cs
@@ -1,19 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-namespace Chinook.ClientModels.Request
-{
-    public class AddTrackToPlaylistModel
-    {
-        [JsonPropertyName("userId")]
-        public string UserId { get; set; }
-
-        [JsonPropertyName("trackId")]
-        public long TrackId { get; set; }
-
-        [JsonPropertyName("playlistId")]
-        public long? PlaylistId { get; set; }
-
-        [JsonPropertyName("newPlaylistName")]
-        public string? NewPlaylistName { get; set; }
-    }
-}
