--- Chinook/ClientModels/Request/RemoveTrackFromPlaylistModel.cs
@@ -1,13 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-namespace Chinook.ClientModels.Request
-{
-    public class RemoveTrackFromPlaylistModel
-    {
-        [JsonPropertyName("trackId")]
-        public long TrackId { get; set; }
-
-        [JsonPropertyName("playlistId")]
-        public long PlaylistId { get; set; }
-    }
-}
