+++ Chinook/ClientServices/API/IApiServices.cs
@@ -3,6 +3,5 @@
     public interface IApiServices
     {
         Task<T> GetDataAsync<T>(string url);
-        Task<TResponse> PostDataAsync<TRequest, TResponse>(string url, TRequest requestData);
     }
 }
