+++ Chinook/Infrastructure/Contracts/Services/ITrackService.cs
@@ -1,5 +1,4 @@
-﻿using Chinook.ClientModels.Request;
-using Chinook.Models;
+﻿using Chinook.Models;
 
 namespace Chinook.Infrastructure.Contracts.Services
 {
@@ -10,21 +9,21 @@ namespace Chinook.Infrastructure.Contracts.Services
         /// </summary>
         /// <param name="artistId"></param>
         /// <returns></returns>
-        Task<List<ClientModels.PlaylistTrack>> GetTracksByArtist(CommonRequestModel requestModel);
+        Task<List<ClientModels.PlaylistTrack>> GetTracksByArtist(long artistId);
 
         /// <summary>
         /// dds a track to the user's favorite playlist
         /// </summary>
         /// <param name="trackId"></param>
         /// <returns></returns>
-        Task<bool> FavoriteTrack(CommonRequestModel requestModel);
+        Task<bool> FavoriteTrack(long trackId);
 
         /// <summary>
         /// Removes a track from the user's favorite playlist
         /// </summary>
         /// <param name="trackId"></param>
         /// <returns></returns>
-        Task<bool> UnFavoriteTrack(CommonRequestModel requestModel);
+        Task<bool> UnFavoriteTrack(long trackId);
 
         /// <summary>
         /// Adds a track to an existing playlist or creates a new playlist and adds the track to it.
@@ -33,7 +32,7 @@ namespace Chinook.Infrastructure.Contracts.Services
         /// <param name="playlistId"></param>
         /// <param name="newPlaylistName"></param>
         /// <returns></returns>
-        Task<string> AddTrackToPlaylist(AddTrackToPlaylistModel requestModel);
+        Task<string> AddTrackToPlaylist(long trackId, long? playlistId, string? newPlaylistName);
 
         /// <summary>
         /// Removes a track from a specified playlist.
@@ -41,7 +40,7 @@ namespace Chinook.Infrastructure.Contracts.Services
         /// <param name="trackId"></param>
         /// <param name="playlistId"></param>
         /// <returns></returns>
-        Task<bool> RemoveTrackFromPlaylist(RemoveTrackFromPlaylistModel requestModel);
+        Task<bool> RemoveTrackFromPlaylist(long trackId, long playlistId);
 
     }
 }
