+++ Chinook/Program.cs
@@ -47,7 +47,6 @@ builder.Services.AddScoped<IAppState,AppState>();
 builder.Services.AddScoped<IApiServices, ApiServices>();
 builder.Services.AddScoped<IClientPlaylistService, ClientPlaylistService>();
 builder.Services.AddScoped<IClientArtistService, ClientArtistService>();
-builder.Services.AddScoped<IClientTrackService, ClientTrackService>();
 builder.Services.AddHttpClient<ClientArtistService>(client =>
 {
