+++ Chinook/Services/TrackService.cs
@@ -1,5 +1,4 @@
-﻿using Chinook.ClientModels.Request;
-using Chinook.Infrastructure.Contracts.Repositories;
+﻿using Chinook.Infrastructure.Contracts.Repositories;
 using Chinook.Infrastructure.Contracts.Services;
 
 namespace Chinook.Services
@@ -7,15 +6,17 @@ namespace Chinook.Services
     public class TrackService : ITrackService
     {
         private readonly ITrackRepository TrackRepository;
+        private readonly IUserService UserService;
 
         /// <summary>
         /// Constructor
         /// </summary>
         /// <param name="repository"></param>
         /// <param name="userService"></param>
-        public TrackService(ITrackRepository repository) 
+        public TrackService(ITrackRepository repository,IUserService userService) 
         {
             TrackRepository = repository;
+            UserService = userService;
         }
 
         /// <summary>
@@ -23,9 +24,10 @@ namespace Chinook.Services
         /// </summary>
         /// <param name="artistId"></param>
         /// <returns></returns>
-        public async Task<List<ClientModels.PlaylistTrack>> GetTracksByArtist(CommonRequestModel requestModel)
+        public async Task<List<ClientModels.PlaylistTrack>> GetTracksByArtist(long artistId)
         {
-            var result = await TrackRepository.GetTracksByArtist(requestModel.Id, requestModel.UserId);
+            var userId = await UserService.GetUserId();
+            var result = await TrackRepository.GetTracksByArtist(artistId,userId);
             return result;
         }
 
@@ -34,9 +36,10 @@ namespace Chinook.Services
         /// </summary>
         /// <param name="trackId"></param>
         /// <returns></returns>
-        public async Task<bool> FavoriteTrack(CommonRequestModel requestModel)
+        public async Task<bool> FavoriteTrack(long trackId)
         {
-            var result = await TrackRepository.AddTrackToFavorite(requestModel.Id, requestModel.UserId);
+            var userId = await UserService.GetUserId();
+            var result = await TrackRepository.AddTrackToFavorite(trackId, userId);
             return result;
         }
 
@@ -45,9 +48,10 @@ namespace Chinook.Services
         /// </summary>
         /// <param name="trackId"></param>
         /// <returns></returns>
-        public async Task<bool> UnFavoriteTrack(CommonRequestModel requestModel)
+        public async Task<bool> UnFavoriteTrack(long trackId)
         {
-            var result = await TrackRepository.RemoveTrackFromFavorite(requestModel.Id, requestModel.UserId);
+            var userId = await UserService.GetUserId();
+            var result = await TrackRepository.RemoveTrackFromFavorite(trackId, userId);
             return result;
         }
 
@@ -58,9 +62,10 @@ namespace Chinook.Services
         /// <param name="playlistId"></param>
         /// <param name="newPlaylistName"></param>
         /// <returns></returns>
-        public async Task<string> AddTrackToPlaylist(AddTrackToPlaylistModel requestModel)
+        public async Task<string> AddTrackToPlaylist(long trackId, long? playlistId, string? newPlaylistName)
         {
-            var result = await TrackRepository.AddTrackToPlaylist(requestModel.TrackId, requestModel.PlaylistId, requestModel.NewPlaylistName, requestModel.UserId);
+            var userId = await UserService.GetUserId();
+            var result = await TrackRepository.AddTrackToPlaylist(trackId, playlistId, newPlaylistName,userId);
 
             return result;
         }
@@ -71,9 +76,9 @@ namespace Chinook.Services
         /// <param name="trackId"></param>
         /// <param name="playlistId"></param>
         /// <returns></returns>
-        public async Task<bool> RemoveTrackFromPlaylist(RemoveTrackFromPlaylistModel requestModel)
+        public async Task<bool> RemoveTrackFromPlaylist(long trackId, long playlistId)
         {
-            var result = await TrackRepository.RemoveTrackFromPlaylist(requestModel.TrackId, requestModel.PlaylistId);
+            var result = await TrackRepository.RemoveTrackFromPlaylist(trackId, playlistId);
             return result;
         }
     }
